server/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
marshmallow==3.20.1
flask-marshmallow==1.2.1
PyMySQL==1.1.0
python-dotenv==1.0.0
email-validator==2.1.0
# Pillow and cloudinary removed temporarily - not compatible with Python 3.14 yet
# You can add them later when needed for document uploads
